Move tests related to commands/tests.py to separate test file
